+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/KotlinHandler.java
@@ -4,8 +4,12 @@
 
 package net.sourceforge.pmd.lang.kotlin;
 
+import java.util.concurrent.ExecutorService;
+
 import net.sourceforge.pmd.lang.AbstractPmdLanguageVersionHandler;
+import net.sourceforge.pmd.lang.LanguagePropertyBundle;
 import net.sourceforge.pmd.lang.ast.Parser;
+import net.sourceforge.pmd.lang.kotlin.ast.InternalApiBridge;
 import net.sourceforge.pmd.lang.kotlin.ast.PmdKotlinParser;
 import net.sourceforge.pmd.lang.rule.xpath.impl.XPathHandler;
 
@@ -14,6 +18,22 @@ public class KotlinHandler extends AbstractPmdLanguageVersionHandler {
 
     private static final XPathHandler XPATH_HANDLER = XPathHandler.noFunctionDefinitions();
 
+    private PmdKotlinParser parser;
+
+    /**
+     * @deprecated Since 7.25.0. Don't create this class directly, use {@link KotlinLanguageModule#getInstance()},
+     *             {@link KotlinLanguageModule#createProcessor(LanguagePropertyBundle)},
+     *             {@link KotlinLanguageProcessor#services()} instead.
+     */
+    @Deprecated
+    public KotlinHandler() {
+        // default constructor - needed for backwards compatibility
+    }
+
+    KotlinHandler(ExecutorService timeoutExecutor) {
+        this.parser = InternalApiBridge.newPmdKotlinParser(timeoutExecutor);
+    }
+
     @Override
     public XPathHandler getXPathHandler() {
         return XPATH_HANDLER;
@@ -21,6 +41,6 @@ public class KotlinHandler extends AbstractPmdLanguageVersionHandler {
 
     @Override
     public Parser getParser() {
-        return new PmdKotlinParser();
+        return parser;
     }
 }